+++ utils.cpp
@@ -1,7 +1,7 @@
 #include "utils.h"
 
-bool utils::is_in_str_array(std::string* a, size_t size, std::string& s){
-    for(size_t i=0; i<size; i++){
+bool utils::is_in_str_array(std::string* a, unsigned short size, std::string& s){
+    for(unsigned short i=0; i<size; i++){
         if(a[i] == s)
             return true;
     }
